+++ addon/globalPlugins/remoteClient/dialogs.py
@@ -121,20 +121,17 @@ class ClientPanel(wx.Panel):
 
 	def handle_certificate_failed(self):
 		try:
-			cert_hash = self.key_connector.last_fail_fingerprint
-				
-			wnd = CertificateUnauthorizedDialog(None, fingerprint=cert_hash)
-			a = wnd.ShowModal()
-			if a == wx.ID_YES:
-				config = configuration.get_config()
-				config['trusted_certs'][self.host.GetValue()]=cert_hash
-				if not config.readonly:
-					config.write()
-			if a != wx.ID_YES and a != wx.ID_NO: return
+			connector = self.key_connector
+			if connector is None or not configuration.trust_certificate(
+				connector.address,
+				connector.last_fail_fingerprint,
+			):
+				log.warning("Unable to automatically trust the server certificate")
+				return
 		except Exception as ex:
 			log.error(ex)
 			return
-		self.key_connector.close()
+		connector.close()
 		self.key_connector = None
 		self.generate_key_command(True)
 
@@ -370,6 +367,18 @@ class OptionsDialog(SettingsPanel):
 		self.autoconnect = wx.CheckBox(self, wx.ID_ANY, label=_("Auto-connect to control server on startup"))
 		self.autoconnect.Bind(wx.EVT_CHECKBOX, self.on_autoconnect)
 		sizer.Add(self.autoconnect)
+		# Translators: A checkbox in add-on options dialog to enable update checks when NVDA starts.
+		self.check_updates = wx.CheckBox(self, wx.ID_ANY, label=_("Check for TeleNVDA updates on startup"))
+		sizer.Add(self.check_updates)
+		# Translators: Select whether update checks use stable or development releases.
+		self.update_channel = wx.RadioBox(
+			self,
+			wx.ID_ANY,
+			label=_("Update channel"),
+			choices=(_("Stable releases"), _("Development releases")),
+			style=wx.RA_VERTICAL,
+		)
+		sizer.Add(self.update_channel)
 		#Translators: Whether or not to use a relay server when autoconnecting
 		self.client_or_server = wx.RadioBox(self, wx.ID_ANY, choices=(_("Use Remote Control Server"), _("Host Control Server")), style=wx.RA_VERTICAL)
 		self.client_or_server.Bind(wx.EVT_RADIOBOX, self.on_client_or_server)
@@ -492,6 +501,9 @@ class OptionsDialog(SettingsPanel):
 
 	def onPanelActivated(self):
 		config = configuration.get_config()
+		updates = config.get('updates', {})
+		self.check_updates.SetValue(updates.get('check_at_startup', True))
+		self.update_channel.SetSelection(1 if updates.get('channel', 'stable') == 'dev' else 0)
 		cs = config['controlserver']
 		self_hosted = cs['self_hosted']
 		connection_type = cs['connection_type']
@@ -590,6 +602,8 @@ class OptionsDialog(SettingsPanel):
 		config['ui']['allow_speech_commands'] = self.speech_commands.GetValue()
 		config['ui']['display_motd_once'] = self.motd_once.GetValue()
 		config['ui']['portcheck'] = self.portcheck.GetValue()
+		config['updates']['check_at_startup'] = self.check_updates.GetValue()
+		config['updates']['channel'] = 'dev' if self.update_channel.GetSelection() == 1 else 'stable'
 		if not configuration.readonly:
 			config.write()